More Major Changes
Still don't have a valid way to define the INVALID_LOCATION struct as a const.

Next steps are to use the move_virtual function to build the network of immovable blocks and effectively immovable blocks, aka "sim_edges" so that we can start building the "rooms" that it needs to analyze.